chore: split in 2 files and adjust public interface
